// src/sync/dto/sync-response.dto.ts
// src/sync/dto/sync-response.dto.ts
import {
  OrderSyncDto,
  StoreSyncDto,
  CashboxSyncDto,
  ItemSyncDto,
  StatSyncDto,
  LogSyncDto,
} from './base-sync.dto';

export class SyncResponseDto {
  now: string; // وقت السيرفر الآن (تستخدمه كـ lastSyncAt جديد)

  orders: OrderSyncDto[];
  stores: StoreSyncDto[];
  cashbox: CashboxSyncDto[];
  items: ItemSyncDto[];
  stats: StatSyncDto[];
  logs: LogSyncDto[];
}